[TASK] Rename `RuleContainer` to `DeclarationList`
It contains property declarations not rules, and describing it as a 'list'
rather than a 'container' seems more apt, as the declarations are ordered.

Changes to use the new interface name in the source code and tests will be
submitted as follow-up PRs.
